+++ packages/zoe/test/test-automaticRefund.js
@@ -6,72 +6,23 @@ import test from 'ava';
 import bundleSource from '@agoric/bundle-source';
 import { E } from '@agoric/eventual-send';
 
-import { makePromiseKit } from '@agoric/promise-kit';
 import { Far } from '@agoric/marshal';
 
-import { assert, details as X } from '@agoric/assert';
 import { evalContractBundle } from '../src/contractFacet/evalContractCode';
 
 import { makeZoe } from '../src/zoeService/zoe';
 
-function makeFakeVatAdmin(testContextSetter = undefined, makeRemote = x => x) {
-  // FakeVatPowers isn't intended to support testing of vat termination, it is
-  // provided to allow unit testing of contracts that call zcf.shutdown()
-  let exitMessage;
-  let hasExited = false;
-  let exitWithFailure;
-  const fakeVatPowers = {
-    exitVat: completion => {
-      exitMessage = completion;
-      hasExited = true;
-      exitWithFailure = false;
-    },
-    exitVatWithFailure: reason => {
-      exitMessage = reason;
-      hasExited = true;
-      exitWithFailure = true;
-    },
-  };
-
-  // This is explicitly intended to be mutable so that
-  // test-only state can be provided from contracts
-  // to their tests.
-  const admin = Far('vatAdmin', {
+function makeFakeVatAdmin() {
+  return Far('vatAdmin', {
     createVat: bundle => {
       return harden({
-        root: makeRemote(
-          E(evalContractBundle(bundle)).buildRootObject(
-            fakeVatPowers,
-            undefined,
-            testContextSetter,
-          ),
-        ),
-        adminNode: Far('adminNode', {
-          done: () => {
-            const kit = makePromiseKit();
-            // Don't trigger Node.js's UnhandledPromiseRejectionWarning.
-            // This does not suppress any error messages.
-            kit.promise.catch(_ => {});
-            return kit.promise;
-          },
-          terminateWithFailure: () => {},
-          adminData: () => {},
-        }),
+        root: E(evalContractBundle(bundle)).buildRootObject(),
       });
     },
-    createVatByName: _name => {
-      assert.fail(X`createVatByName not supported in fake mode`);
-    },
   });
-  const vatAdminState = {
-    getExitMessage: () => exitMessage,
-    getHasExited: () => hasExited,
-    getExitWithFailure: () => exitWithFailure,
-  };
-  return { admin, vatAdminState };
 }
 
-const fakeVatAdmin = makeFakeVatAdmin().admin;
+const fakeVatAdmin = makeFakeVatAdmin();
 
 const automaticRefundRoot = `${__dirname}/../src/contracts/automaticRefund`;
